refactor: extract proxy provider management to extra class
