+++ src/index.ts
@@ -8,3 +8,4 @@ export * from './lib/cls.service';
 export * from './lib/cls.decorators';
 export * from './lib/cls.interfaces';
 export { Terminal } from './types/terminal.type';
+export { InjectableProxy } from './lib/proxy-provider';
