+++ src/lib/cls.constants.ts
@@ -5,4 +5,3 @@ export const CLS_MODULE_OPTIONS = 'ClsModuleOptions';
 export const CLS_MIDDLEWARE_OPTIONS = 'ClsMiddlewareOptions';
 export const CLS_GUARD_OPTIONS = 'ClsGuardOptions';
 export const CLS_INTERCEPTOR_OPTIONS = 'ClsInterceptorOptions';
-export const CLS_PROXY_METADATA_KEY = '__cls_proxy__';
