+++ src/lib/proxy-provider/proxy-provider.constants.ts
@@ -1,0 +1,1 @@
+export const CLS_PROXY_METADATA_KEY = '__cls_proxy__';
